Core/API: Add IntervalBar
IntervalBar works like Bar, but gets two times instead of one. The first time
indicates the earliest occurrence of the event, the second one represents the
latest occurrence. The timers works as usual, with one difference: when the
first timer is reached (earliest occurrence), the timer stays at zero and does
not disappear. Then, the bar fades progressively for the remaining time until
the latest occurrence is reached. This should give users a clear indication if
an event, represented by the bar, already happened.